Add ability to make reports and dashboards public
Public means they are available to anonymous users.

There is a similar feature that allows to publish reports and dashboards
to a company's public space, which is only available to authenticated
users.
To avoid confusion, this has been reworded to "shared space"

Add new dependency on ngclipboard, to copy public reports and dashboards
URLs into the user's clipboard.